Refactored package structure of api subpackage, improved testability of various classes by passing in dependencies as parameters
