+++ schedoscope-core/pom.xml
@@ -72,6 +72,16 @@
 			<scope>test</scope>
 		</dependency>
 		<dependency>
+			<groupId>ch.qos.logback</groupId>
+			<artifactId>logback-core</artifactId>
+			<version>1.1.2</version>
+		</dependency>
+		<dependency>
+			<groupId>ch.qos.logback</groupId>
+			<artifactId>logback-classic</artifactId>
+			<version>1.1.2</version>
+		</dependency>
+		<dependency>
 			<groupId>commons-io</groupId>
 			<artifactId>commons-io</artifactId>
 			<version>2.4</version>
@@ -136,21 +146,6 @@
 			<groupId>io.spray</groupId>
 			<artifactId>spray-routing_${scala.version}</artifactId>
 			<version>1.3.2</version>
-		</dependency>
-		<dependency>
-			<groupId>io.kamon</groupId>
-			<artifactId>kamon-core_${scala.version}</artifactId>
-			<version>0.5.2</version>
-		</dependency>
-		<dependency>
-			<groupId>io.kamon</groupId>
-			<artifactId>kamon-statsd_${scala.version}</artifactId>
-			<version>0.5.2</version>
-		</dependency>
-		<dependency>
-			<groupId>io.kamon</groupId>
-			<artifactId>kamon-akka_${scala.version}</artifactId>
-			<version>0.5.2</version>
 		</dependency>
 		<dependency>
 			<groupId>com.github.scopt</groupId>
@@ -286,15 +281,6 @@
 
 	<build>
 		<plugins>
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-compiler-plugin</artifactId>
-				<version>3.3</version>
-				<configuration>
-					<source>1.7</source>
-					<target>1.7</target>
-				</configuration>
-			</plugin>
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
