--- schedoscope-core/src/main/resources/graph.html
@@ -1,166 +1,0 @@
-<!DOCTYPE html>
-<meta charset="utf-8">
-<style>
-.node circle {
-	stroke: #ccc;
-	stroke-width: 0.5px;
-}
-
-.node {
-	font: 10px sans-serif;
-}
-
-.link {
-	fill: none;
-	stroke: #ccc;
-	stroke-width: 0.5px;
-}
-</style>
-<body>
-	<script src="http://d3js.org/d3.v3.min.js"></script>
-	<script>
-
-var diameter = 1400;
-
-var colors = {};
-colors['root'] = "black";
-colors['table'] = "brown";
-colors['dummy'] = "white";
-colors['receive'] = "grey";
-colors['materialized'] = "green";
-colors['waiting'] = "yellow";
-colors['transforming'] = "orange";
-colors['error'] = "red";
-colors['nodata'] = "white";
-
-var viewUrlPath = window.location.pathname.split("/").slice(2).join("/");
-
-var query = window.location.search.substring(1);
-var vars = query.split("&");
-var filter = "";
-var initialRoot = "ROOT";
-var exclude = undefined;
-vars.forEach(function(v) {
-    p=v.split("=");
-    if (p[0] == "filter") {
-        filter = "&filter=.*" + p[1] + ".*";
-    }
-    if (p[0] == "root") {
-        initialRoot = p[1];
-    }
-    if (p[0] == "exclude") {
-        exclude = new RegExp(".*" + p[1] + ".*");
-    }
-
-})
-
-d3.json("/views/" + viewUrlPath + "?dependencies=true" + filter, function(error, views) {
-  // build lookup view-url -> view
-  var viewsMap = {};
-  var hasParentCount = {};
-  views.views.forEach(function(view) {
-    if ((!exclude || !exclude.test(view.view)) && view.status != "nodata" && view.status != "receive") {
-      viewsMap[view.view] = view;
-      hasParentCount[view.view] = 0;
-    }
-  });
-
-  // reformat to tree-conform syntax
-  views.views.forEach(function(view) {
-    var par = viewsMap[view.view];
-    if (par) { 
-      par.name = par.view;
-      delete par.view;
-      view.dependencies.forEach(function(dep) {
-        var child = viewsMap[dep.view];
-        if (child) {
-          hasParentCount[dep.view] += 1;
-          if (!par.children) par.children = []; 
-          par.children.push(child);
-        }
-      });
-      delete par.dependencies
-    }
-  });
-
-  // add top-level nodes for databases / root node
-  var rootNode = {name:"ROOT", status:"root", children: []};
-  viewsMap["ROOT"] = rootNode;
-  for (var key in hasParentCount) {
-    if (hasParentCount[key] == 0) {
-      var tableName = key.split("/").slice(0,2).join("/");
-      if (!(tableName in viewsMap)) {
-        var tableNode = {name: tableName, status:"table", children: []};
-        rootNode.children.push(tableNode);
-        viewsMap[tableName] = tableNode;
-      }
-      viewsMap[tableName].children.push(viewsMap[key]);
-    }
-  }
-  if (viewsMap["ROOT"].children.length == 1) {
-    viewsMap["ROOT"] = viewsMap["ROOT"].children[0];
-  }
-
-d3.select("body").append("p").attr("id", "nav").append("a").attr("href", "#").on("click", function(d) {draw("ROOT")}).text("root");
-
-draw(initialRoot);
-
-function draw(rootName) {
-
-var tree = d3.layout.tree()
-    .size([360, diameter / 2 - 120])
-    .separation(function(a, b) { return (a.parent == b.parent ? 1 : 2) / a.depth; });
-
-var diagonal = d3.svg.diagonal.radial()
-    .projection(function(d) { return [d.y, d.x / 180 * Math.PI]; });
-
-d3.select("body").selectAll("svg").remove();
-
-var svg = d3.select("body").append("svg")
-    .attr("width", diameter)
-    .attr("height", diameter - 150)
-  .append("g")
-    .attr("transform", "translate(" + diameter / 2 + "," + diameter / 2 + ")");
-
-  var root = viewsMap[rootName];
-  //alert(JSON.stringify(root));
-
-  // if root has 1 child, the radial tree layout is broken -> add dummy node in such a case
-  var fixed = false;
-  if (root.children.length <= 1) {
-    root.children.push({name:"dummy",status:"dummy"});
-    fixed = true;
-  }
-
-  var nodes = tree.nodes(root),
-      links = tree.links(nodes);
-
-  var link = svg.selectAll(".link")
-      .data(links)
-    .enter().append("path")
-      .attr("class", "link")
-      .attr("d", diagonal);
-
-  var node = svg.selectAll(".node")
-      .data(nodes)
-    .enter().append("g")
-      .attr("class", "node")
-      .attr("transform", function(d) { return "rotate(" + (d.x - 90) + ")translate(" + d.y + ")"; })
-      .on("click", function(d) { draw(d.name) } );
-
-  node.append("circle")
-      .attr("r", 3.5)
-      .style("fill", function(d) { return colors[d.status] });
-
-  node.append("title")
-      .text(function(d) { return d.name + ": " + d.status; });
-
-  if (fixed) root.children.pop();
-}
-
-});
-
-
-d3.select(self.frameElement).style("height", diameter - 150 + "px");
-
-</script>
