+++ schedoscope-core/src/main/scala/org/schedoscope/Settings.scala
@@ -42,7 +42,7 @@ import org.schedoscope.scheduler.driver.Driver
 import org.schedoscope.dsl.views.ViewUrlParser.ParsedViewAugmentor
 
 /**
- * The Settings class stores all settings for Schedoscope. It is a singelton accessible through
+ * The Settings class keeps all settings for Schedoscope. It is a singelton accessible through
  * the companion object's apply() method.
  *
  * Configuration is based on TypeSafe config. For an explanation of the different
@@ -50,18 +50,26 @@ import org.schedoscope.dsl.views.ViewUrlParser.ParsedViewAugmentor
  *
  */
 class SettingsImpl(val config: Config) extends Extension {
-  val system = Settings.actorSystem
 
   private val driverSettings: HashMap[String, DriverSettings] = HashMap[String, DriverSettings]()
 
+  /**
+   * The configured Schedoscope environment
+   */
   lazy val env = config.getString("schedoscope.app.environment")
 
+  /**
+   * The configured earliest day for DateParamterization logic
+   */
   lazy val earliestDay = {
     val conf = config.getString("schedoscope.scheduler.earliestDay")
     val Array(year, month, day) = conf.split("-")
     DateParameterizationUtils.parametersToDay(p(year), p(month), p(day))
   }
 
+  /**
+   * The configured latest, i.e., current day for DateParamterization logic
+   */
   lazy val latestDay = {
     val conf = config.getString("schedoscope.scheduler.latestDay")
     if (conf == "now") {
@@ -77,34 +85,76 @@ class SettingsImpl(val config: Config) extends Extension {
     }
   }
 
+  /**
+   * The configured timeout for Schedoscope web service calls.
+   */
   lazy val webserviceTimeout =
     Duration.create(config.getDuration("schedoscope.webservice.timeout", TimeUnit.MILLISECONDS),
       TimeUnit.MILLISECONDS)
 
+  /**
+   * The configured host of the Schedoscope web service (for calling via the SchedoscopeRestClient).
+   */
   lazy val host = config.getString("schedoscope.webservice.host")
 
+  /**
+   * The configured post of the Schedoscope web service (for both web service and client)
+   */
   lazy val port = config.getInt("schedoscope.webservice.port")
 
+  /**
+   * Configured directory for storing static web service resource data for Schedoscope web service
+   */
   lazy val webResourcesDirectory = config.getString("schedoscope.webservice.resourceDirectory")
 
+  /**
+   * Configured actors for Schedoscope web service.
+   */
   lazy val restApiConcurrency = config.getInt("schedoscope.webservice.concurrency")
 
+  /**
+   * Configured JDBC URL to the Hive server
+   */
   lazy val jdbcUrl = config.getString("schedoscope.metastore.jdbcUrl")
 
+  /**
+   * Configured kerberos principal.
+   */
   lazy val kerberosPrincipal = config.getString("schedoscope.kerberos.principal")
 
+  /**
+   * Configured Thrift metastore URI.
+   */
   lazy val metastoreUri = config.getString("schedoscope.metastore.metastoreUri")
 
+  /**
+   * Configured view augmentor class for postprocessing of views after URL parsing
+   */
   lazy val parsedViewAugmentorClass = config.getString("schedoscope.app.parsedViewAugmentorClass")
 
-  def viewAugmentor = Class.forName(parsedViewAugmentorClass).newInstance().asInstanceOf[ParsedViewAugmentor]
+  /**
+   * An instance of the view augmentor class
+   */
+  lazy val viewAugmentor = Class.forName(parsedViewAugmentorClass).newInstance().asInstanceOf[ParsedViewAugmentor]
 
+  /**
+   * The subconfigurations of the configured transformation types.
+   */
   lazy val availableTransformations = config.getObject("schedoscope.transformations")
 
+  /**
+   * A suitable hadoop config to use
+   */
   lazy val hadoopConf = new Configuration(true)
 
+  /**
+   * Configuration trigger whether versioning transformation is enabled.
+   */
   lazy val transformationVersioning = config.getBoolean("schedoscope.versioning.transformations")
 
+  /**
+   * Return the resource manager or job tracker equivalent
+   */
   lazy val jobTrackerOrResourceManager = {
     val yarnConf = new YarnConfiguration(hadoopConf)
     if (yarnConf.get("yarn.resourcemanager.address") == null)
@@ -113,23 +163,62 @@ class SettingsImpl(val config: Config) extends Extension {
       yarnConf.get("yarn.resourcemanager.address")
   }
 
+  /**
+   * The configured HDFS namenode. The Hadoop default takes precendence.
+   */
   lazy val nameNode = if (hadoopConf.get("fs.defaultFS") == null)
     config.getString("schedoscope.hadoop.nameNode")
   else
     hadoopConf.get("fs.defaultFS")
 
+  /**
+   * The configured timeout for filesystem operations.
+   */
   lazy val filesystemTimeout = getDriverSettings("filesystem").timeout
+
+  /**
+   * The configured timeout for schema / metastore operations.
+   */
   lazy val schemaTimeout = Duration.create(config.getDuration("schedoscope.scheduler.timeouts.schema", TimeUnit.SECONDS), TimeUnit.SECONDS)
+
+  /**
+   * The configured timeout for collecting status lists (view / transformation)
+   */
   lazy val statusListAggregationTimeout = Duration.create(config.getDuration("schedoscope.scheduler.timeouts.statusListAggregation", TimeUnit.SECONDS), TimeUnit.SECONDS)
+
+  /**
+   * The configured timout for responses of the view manager.
+   */
   lazy val viewManagerResponseTimeout = Duration.create(config.getDuration("schedoscope.scheduler.timeouts.viewManagerResponse", TimeUnit.SECONDS), TimeUnit.SECONDS)
+
+  /**
+   * Timeout for completion of materialization commands
+   */
   lazy val completitionTimeout = Duration.create(config.getDuration("schedoscope.scheduler.timeouts.completion", TimeUnit.SECONDS), TimeUnit.SECONDS)
 
+  /**
+   * The configured number of retries before a view enters failed state.
+   */
   lazy val retries = config.getInt("schedoscope.action.retry")
 
+  /**
+   * Number of parallel threads to access the metastore
+   */
   lazy val metastoreConcurrency = config.getInt("schedoscope.metastore.concurrency")
+
+  /**
+   * Number of partitions to write in batch to the metatatore
+   */
   lazy val metastoreWriteBatchSize = config.getInt("schedoscope.metastore.writeBatchSize")
+
+  /**
+   * Number of partitions to read in batch to the metatatore
+   */
   lazy val metastoreReadBatchSize = config.getInt("schedoscope.metastore.readBatchSize")
 
+  /**
+   * A user group information object ready to use for kerberized interactions.
+   */
   lazy val userGroupInformation = {
     UserGroupInformation.setConfiguration(hadoopConf)
     val ugi = UserGroupInformation.getCurrentUser()
@@ -188,9 +277,10 @@ class SettingsImpl(val config: Config) extends Extension {
 
 }
 
+/**
+ * Companion object for settings
+ */
 object Settings extends ExtensionId[SettingsImpl] with ExtensionIdProvider {
-  val actorSystem = ActorSystem("schedoscope")
-
   override def lookup = Settings
 
   override def createExtension(system: ExtendedActorSystem) =
@@ -199,18 +289,47 @@ object Settings extends ExtensionId[SettingsImpl] with ExtensionIdProvider {
   override def get(system: ActorSystem): SettingsImpl = super.get(system)
 
   def apply() = {
-    super.apply(actorSystem)
+    super.apply(Schedoscope.actorSystem)
   }
 }
 
+/**
+ * Accessor for transformation driver-related settings.
+ */
 class DriverSettings(val config: Config, val name: String) {
+  /**
+   * Location where to put transformation-driver related resources into HDFS upon Schedoscope start.
+   */
   lazy val location = Settings().nameNode + config.getString("location") + Settings().env + "/"
+
+  /**
+   * Comma-separated list of additional jars to put in HDFS location upon Schedoscope start.
+   */
   lazy val libDirectory = config.getString("libDirectory")
+
+  /**
+   * Number of parallel drivers.
+   */
   lazy val concurrency = config.getInt("concurrency")
+
+  /**
+   * Do the driver respective resource need unpacking
+   */
   lazy val unpack = config.getBoolean("unpack")
+
+  /**
+   * URL for the driver to access the service ultimately executing the transformation, e.g., the hive server.
+   */
   lazy val url = config.getString("url")
+
+  /**
+   * Timeout for transformation of this type.
+   */
   lazy val timeout = Duration.create(config.getDuration("timeout", TimeUnit.SECONDS), TimeUnit.SECONDS)
 
+  /**
+   * List of jars to upload to HDFS.
+   */
   lazy val libJars = {
     val fromLibDir = libDirectory
       .split(",")
@@ -233,6 +352,9 @@ class DriverSettings(val config: Config, val name: String) {
     (fromLibDir ++ fromClasspath).toList
   }
 
+  /**
+   * Paths of jars in HDFS.
+   */
   lazy val libJarsHdfs = {
     if (unpack)
       List[String]()
@@ -241,5 +363,8 @@ class DriverSettings(val config: Config, val name: String) {
     }
   }
 
+  /**
+   * Configured driver run completion handlers.
+   */
   lazy val driverRunCompletionHandlers = config.getStringList("driverRunCompletionHandlers").toList
 }
