+++ schedoscope-core/src/main/scala/org/schedoscope/scheduler/actors/MetadataLoggerActor.scala
@@ -13,17 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.schedoscope.scheduler
+package org.schedoscope.scheduler.actors
 
-import scala.collection.mutable.HashMap
 import org.schedoscope.schema.SchemaManager
 import akka.actor.Actor
 import akka.actor.Props
 import akka.actor.actorRef2Scala
 import akka.event.Logging
 import akka.event.LoggingReceive
-import akka.routing.SmallestMailboxRoutingLogic
-import akka.routing.Router
 import org.schedoscope.scheduler.messages._
 
 /**
