+++ schedoscope-core/src/main/scala/org/schedoscope/scheduler/actors/SchemaActor.scala
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.schedoscope.scheduler
+package org.schedoscope.scheduler.actors
 
-import scala.collection.mutable.HashMap
 import org.schedoscope.schema.SchemaManager
 import org.schedoscope.scheduler.messages._
 import akka.actor.Actor
@@ -23,8 +22,6 @@ import akka.actor.Props
 import akka.actor.actorRef2Scala
 import akka.event.Logging
 import akka.event.LoggingReceive
-import akka.routing.SmallestMailboxRoutingLogic
-import akka.routing.Router
 
 /**
  * Schema actors are responsible for creating tables and partitions in the metastore.
