+++ schedoscope-tutorial/pom.xml
@@ -15,16 +15,6 @@
 			<scope>test</scope>
 		</dependency>
 		<dependency>
-			<groupId>ch.qos.logback</groupId>
-			<artifactId>logback-core</artifactId>
-			<version>1.1.2</version>
-		</dependency>
-		<dependency>
-			<groupId>ch.qos.logback</groupId>
-			<artifactId>logback-classic</artifactId>
-			<version>1.1.2</version>
-		</dependency>
-		<dependency>
 			<groupId>schedoscope</groupId>
 			<artifactId>schedoscope-core</artifactId>
 			<version>0.3.0-SNAPSHOT</version>
@@ -300,7 +290,7 @@
 					</execution>
 				</executions>
 				<configuration>
-					<mainClass>org.schedoscope.scheduler.api.SchedoscopeRestService</mainClass>
+					<mainClass>org.schedoscope.scheduler.rest.server.SchedoscopeRestService</mainClass>
 					<arguments>
 						<argument>--shell</argument>
 					</arguments>
